backend/requirements.txt
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-decouple
Pillow
django-email-verification
psycopg2-binary
